vrclient: Revert "vrclient: HACK: Fixup digital action 'bChange' parameter on 32 bit."
By removing manual handling of IVRInput::GetDigitalActionData().

CW-Bug-Id: #19067
CW-Bug-Id: #24440
