+++ vrclient_x64/Makefile.in
@@ -11,7 +11,6 @@ SOURCES = \
 	vrclient_generated.c \
 	vrclient_main.c \
 	vrcompositor_manual.c \
-	vrinput_manual.c \
 	vroverlay_manual.c \
 	vrsystem_manual.c \
 	vrrendermodels_manual.c \
