+++ vrclient_x64/gen_wrapper.py
@@ -251,7 +251,6 @@ MANUAL_METHODS = {
     "IVRRenderModels_FreeRenderModel": lambda ver, abi: ver > 1,
     "IVRMailbox_undoc3": lambda ver, abi: abi == 'u',
     "IVROverlay_SetOverlayTexture": True,
-    "IVRInput_GetDigitalActionData": lambda ver, abi: abi == 'w' and ver > 3,
     "IVRTrackedCamera_GetVideoStreamFrame": True,
 }
