+++ vrclient_x64/winIVRInput.c
@@ -454,7 +454,24 @@ uint32_t __thiscall winIVRInput_IVRInput_004_UpdateActionState(struct w_steam_if
     return params._ret;
 }
 
-extern uint32_t __thiscall winIVRInput_IVRInput_004_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice);
+uint32_t __thiscall winIVRInput_IVRInput_004_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
+{
+    w_InputDigitalActionData_t w_pActionData;
+    struct IVRInput_IVRInput_004_GetDigitalActionData_params params =
+    {
+        .linux_side = _this->u_iface,
+        .action = action,
+        .pActionData = pActionData ? &w_pActionData : NULL,
+        .unActionDataSize = unActionDataSize,
+        .ulRestrictToDevice = ulRestrictToDevice,
+    };
+    TRACE("%p\n", _this);
+    unActionDataSize = min( unActionDataSize, sizeof(w_pActionData) );
+    if (pActionData) memcpy( &w_pActionData, pActionData, unActionDataSize );
+    VRCLIENT_CALL( IVRInput_IVRInput_004_GetDigitalActionData, &params );
+    if (pActionData) memcpy( pActionData, &w_pActionData, unActionDataSize );
+    return params._ret;
+}
 
 uint32_t __thiscall winIVRInput_IVRInput_004_GetAnalogActionData(struct w_steam_iface *_this, uint64_t action, w_InputAnalogActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
 {
@@ -832,7 +849,24 @@ uint32_t __thiscall winIVRInput_IVRInput_005_UpdateActionState(struct w_steam_if
     return params._ret;
 }
 
-extern uint32_t __thiscall winIVRInput_IVRInput_005_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice);
+uint32_t __thiscall winIVRInput_IVRInput_005_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
+{
+    w_InputDigitalActionData_t w_pActionData;
+    struct IVRInput_IVRInput_005_GetDigitalActionData_params params =
+    {
+        .linux_side = _this->u_iface,
+        .action = action,
+        .pActionData = pActionData ? &w_pActionData : NULL,
+        .unActionDataSize = unActionDataSize,
+        .ulRestrictToDevice = ulRestrictToDevice,
+    };
+    TRACE("%p\n", _this);
+    unActionDataSize = min( unActionDataSize, sizeof(w_pActionData) );
+    if (pActionData) memcpy( &w_pActionData, pActionData, unActionDataSize );
+    VRCLIENT_CALL( IVRInput_IVRInput_005_GetDigitalActionData, &params );
+    if (pActionData) memcpy( pActionData, &w_pActionData, unActionDataSize );
+    return params._ret;
+}
 
 uint32_t __thiscall winIVRInput_IVRInput_005_GetAnalogActionData(struct w_steam_iface *_this, uint64_t action, w_InputAnalogActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
 {
@@ -1317,7 +1351,24 @@ uint32_t __thiscall winIVRInput_IVRInput_006_UpdateActionState(struct w_steam_if
     return params._ret;
 }
 
-extern uint32_t __thiscall winIVRInput_IVRInput_006_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice);
+uint32_t __thiscall winIVRInput_IVRInput_006_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
+{
+    w_InputDigitalActionData_t w_pActionData;
+    struct IVRInput_IVRInput_006_GetDigitalActionData_params params =
+    {
+        .linux_side = _this->u_iface,
+        .action = action,
+        .pActionData = pActionData ? &w_pActionData : NULL,
+        .unActionDataSize = unActionDataSize,
+        .ulRestrictToDevice = ulRestrictToDevice,
+    };
+    TRACE("%p\n", _this);
+    unActionDataSize = min( unActionDataSize, sizeof(w_pActionData) );
+    if (pActionData) memcpy( &w_pActionData, pActionData, unActionDataSize );
+    VRCLIENT_CALL( IVRInput_IVRInput_006_GetDigitalActionData, &params );
+    if (pActionData) memcpy( pActionData, &w_pActionData, unActionDataSize );
+    return params._ret;
+}
 
 uint32_t __thiscall winIVRInput_IVRInput_006_GetAnalogActionData(struct w_steam_iface *_this, uint64_t action, w_InputAnalogActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
 {
@@ -1827,7 +1878,24 @@ uint32_t __thiscall winIVRInput_IVRInput_007_UpdateActionState(struct w_steam_if
     return params._ret;
 }
 
-extern uint32_t __thiscall winIVRInput_IVRInput_007_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice);
+uint32_t __thiscall winIVRInput_IVRInput_007_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
+{
+    w_InputDigitalActionData_t w_pActionData;
+    struct IVRInput_IVRInput_007_GetDigitalActionData_params params =
+    {
+        .linux_side = _this->u_iface,
+        .action = action,
+        .pActionData = pActionData ? &w_pActionData : NULL,
+        .unActionDataSize = unActionDataSize,
+        .ulRestrictToDevice = ulRestrictToDevice,
+    };
+    TRACE("%p\n", _this);
+    unActionDataSize = min( unActionDataSize, sizeof(w_pActionData) );
+    if (pActionData) memcpy( &w_pActionData, pActionData, unActionDataSize );
+    VRCLIENT_CALL( IVRInput_IVRInput_007_GetDigitalActionData, &params );
+    if (pActionData) memcpy( pActionData, &w_pActionData, unActionDataSize );
+    return params._ret;
+}
 
 uint32_t __thiscall winIVRInput_IVRInput_007_GetAnalogActionData(struct w_steam_iface *_this, uint64_t action, w_InputAnalogActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
 {
@@ -2376,7 +2444,24 @@ uint32_t __thiscall winIVRInput_IVRInput_010_UpdateActionState(struct w_steam_if
     return params._ret;
 }
 
-extern uint32_t __thiscall winIVRInput_IVRInput_010_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice);
+uint32_t __thiscall winIVRInput_IVRInput_010_GetDigitalActionData(struct w_steam_iface *_this, uint64_t action, w_InputDigitalActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
+{
+    w_InputDigitalActionData_t w_pActionData;
+    struct IVRInput_IVRInput_010_GetDigitalActionData_params params =
+    {
+        .linux_side = _this->u_iface,
+        .action = action,
+        .pActionData = pActionData ? &w_pActionData : NULL,
+        .unActionDataSize = unActionDataSize,
+        .ulRestrictToDevice = ulRestrictToDevice,
+    };
+    TRACE("%p\n", _this);
+    unActionDataSize = min( unActionDataSize, sizeof(w_pActionData) );
+    if (pActionData) memcpy( &w_pActionData, pActionData, unActionDataSize );
+    VRCLIENT_CALL( IVRInput_IVRInput_010_GetDigitalActionData, &params );
+    if (pActionData) memcpy( pActionData, &w_pActionData, unActionDataSize );
+    return params._ret;
+}
 
 uint32_t __thiscall winIVRInput_IVRInput_010_GetAnalogActionData(struct w_steam_iface *_this, uint64_t action, w_InputAnalogActionData_t *pActionData, uint32_t unActionDataSize, uint64_t ulRestrictToDevice)
 {
